+++ sdk/dynamicBrcode/index.js
@@ -1,0 +1,7 @@
+const dynamicBrcode = require('./dynamicBrcode.js');
+
+exports.create = dynamicBrcode.create;
+exports.query = dynamicBrcode.query;
+exports.get = dynamicBrcode.get;
+exports.page = dynamicBrcode.page;
+exports.DynamicBrcode = dynamicBrcode.DynamicBrcode;